ruido_cli/requirements.txt
requests==2.32.3
pandas==2.2.2
PyYAML==6.0.2
python-dotenv==1.0.1
click==8.1.7
